MdeModulePkg/SdMmcPciHcDxe: Support override for SD 1.8v signaling switch
Some platforms (e.g. Raspberry Pi) have SDHCI implementations where the
internal signaling voltage control isn't wired up and therefore needs to
be set externally, in a non-standard way.

Power cycling the card to get it out of 1.8V is also kind of necessary,
but for now it can be done in the Reset/InitHost phase override.

Signed-off-by: Mario Bălănică <mariobalanica02@gmail.com>